+++ test/symmetries/group_action/test_group_action_lookup.cpp
@@ -9,16 +9,16 @@
 
 using namespace hydra;
 
-template <class bit_t> void test_permutation_group_lookup(int n_sites) {
+template <class bit_t> void test_permutation_group_lookup(int64_t n_sites) {
   using combinatorics::Subsets;
 
   // test cyclic group
   std::vector<Permutation> permutation_array;
-  for (int sym = 0; sym < n_sites; ++sym) {
+  for (int64_t sym = 0; sym < n_sites; ++sym) {
 
-    std::vector<int> pv;
-    for (int site = 0; site < n_sites; ++site) {
-      int newsite = (site + sym) % n_sites;
+    std::vector<int64_t> pv;
+    for (int64_t site = 0; site < n_sites; ++site) {
+      int64_t newsite = (site + sym) % n_sites;
       pv.push_back(newsite);
     }
     permutation_array.push_back(Permutation(pv));
@@ -28,7 +28,7 @@ template <class bit_t> void test_permutation_group_lookup(int n_sites) {
   auto lookup = GroupActionLookup<bit_t>(perm_group);
 
   // Check whether representative is smallest in orbit
-  for (int sym = 0; sym < perm_group.size(); ++sym)
+  for (int64_t sym = 0; sym < perm_group.size(); ++sym)
     for (bit_t state : Subsets<bit_t>(n_sites)) {
 
       auto s1 = action.apply(sym, state);
@@ -39,7 +39,7 @@ template <class bit_t> void test_permutation_group_lookup(int n_sites) {
 
 TEST_CASE("GroupActionLookup", "[symmetries]") {
   hydra::Log.out("PermutationGroupLookup <-> GroupAction comparison");
-  for (int n_sites = 1; n_sites < 6; ++n_sites) {
+  for (int64_t n_sites = 1; n_sites < 6; ++n_sites) {
     test_permutation_group_lookup<uint16_t>(n_sites);
     test_permutation_group_lookup<uint32_t>(n_sites);
     test_permutation_group_lookup<uint64_t>(n_sites);
